feat: simplify and upgrade
