+++ pyproject.toml
@@ -19,5 +19,6 @@ test = [
 dev = [
     "pytest>=8.4.1",
     "pytest-asyncio>=1.1.0",
+    "ruff>=0.15.1",
 ]
