+++ test/test_renamer.py
@@ -2,16 +2,19 @@
 """
 Unit tests for jellyfin-renamer using pytest framework.
 """
+
 import shutil
 import tempfile
 from pathlib import Path
 
 import pytest
 
-from core.movie_organizer import organize_movies
-from core.movie_parser import parse_movie_info
-from core.tv_organizer import organize_tv_shows
-from core.tv_parser import detect_content_type, parse_tv_info
+from core.organizer import organize
+from core.parser import (
+    get_base_filename_without_ext,
+    get_real_extension,
+    parse_media_info,
+)
 
 
 @pytest.mark.parametrize(
@@ -33,7 +36,7 @@ from core.tv_parser import detect_content_type, parse_tv_info
             "1080p",
             None,
             None,
-        ),  # guessit truncates long titles
+        ),
         (
             "Avatar (2009) 1080p - version1.mkv",
             "Avatar",
@@ -41,7 +44,7 @@ from core.tv_parser import detect_content_type, parse_tv_info
             "1080p",
             None,
             None,
-        ),  # guessit doesn't detect "version1" as extra_type
+        ),
         (
             "The Hobbit - An Unexpected Journey (2012) 1080p - part1.mkv",
             "The Hobbit",
@@ -49,7 +52,7 @@ from core.tv_parser import detect_content_type, parse_tv_info
             "1080p",
             1,
             None,
-        ),  # guessit detects part correctly
+        ),
         (
             "The Matrix (1999) 1080p - Trailer.mkv",
             "The Matrix",
@@ -77,21 +80,23 @@ def test_movie_parsing(
     expected_extra_type,
 ):
     """Test movie filename parsing with various patterns."""
-    title, year, resolution, part, extra_type = parse_movie_info(filename)
+    info = parse_media_info(filename, "movies")
 
-    assert (
-        title == expected_title
-    ), f"Title mismatch: got '{title}', expected '{expected_title}'"
-    assert year == expected_year, f"Year mismatch: got {year}, expected {expected_year}"
-    assert (
-        resolution == expected_resolution
-    ), f"Resolution mismatch: got '{resolution}', expected '{expected_resolution}'"
-    assert (
-        part == expected_part
-    ), f"Part mismatch: got '{part}', expected '{expected_part}'"
-    assert (
-        extra_type == expected_extra_type
-    ), f"Extra type mismatch: got '{extra_type}', expected '{expected_extra_type}'"
+    assert info["title"] == expected_title, (
+        f"Title mismatch: got '{info['title']}', expected '{expected_title}'"
+    )
+    assert info["year"] == expected_year, (
+        f"Year mismatch: got {info['year']}, expected {expected_year}"
+    )
+    assert info["resolution"] == expected_resolution, (
+        f"Resolution mismatch: got '{info['resolution']}', expected '{expected_resolution}'"
+    )
+    assert info["part"] == expected_part, (
+        f"Part mismatch: got {info['part']}, expected {expected_part}"
+    )
+    assert info["extra_type"] == expected_extra_type, (
+        f"Extra type mismatch: got {info['extra_type']}, expected {expected_extra_type}"
+    )
 
 
 @pytest.mark.parametrize(
@@ -166,7 +171,7 @@ def test_movie_parsing(
             "1080p",
             None,
             None,
-        ),  # guessit truncates long series names
+        ),
         (
             "Mr.Robot.S01E01.eps1.0_hellofriend.mov.1080p.DTS-HD.MA.5.1.AVC.REMUX.mkv",
             "Mr Robot",
@@ -190,47 +195,29 @@ def test_tv_parsing(
     expected_extra_type,
 ):
     """Test TV show filename parsing with various patterns."""
-    series, season, episodes, year, resolution, part, extra_type = parse_tv_info(
-        filename
+    info = parse_media_info(filename, "tv")
+
+    assert info["title"] == expected_series, (
+        f"Series mismatch: got '{info['title']}', expected '{expected_series}'"
     )
-
-    assert (
-        series == expected_series
-    ), f"Series mismatch: got '{series}', expected '{expected_series}'"
-    assert (
-        season == expected_season
-    ), f"Season mismatch: got {season}, expected {expected_season}"
-    assert (
-        episodes == expected_episodes
-    ), f"Episodes mismatch: got {episodes}, expected {expected_episodes}"
-    assert year == expected_year, f"Year mismatch: got {year}, expected {expected_year}"
-    assert (
-        resolution == expected_resolution
-    ), f"Resolution mismatch: got '{resolution}', expected '{expected_resolution}'"
-    assert (
-        part == expected_part
-    ), f"Part mismatch: got '{part}', expected '{expected_part}'"
-    assert (
-        extra_type == expected_extra_type
-    ), f"Extra type mismatch: got '{extra_type}', expected '{expected_extra_type}'"
-
-
-@pytest.mark.parametrize(
-    "filename,expected_type",
-    [
-        ("The.Matrix.1999.1080p.BluRay.x264.mkv", "movie"),
-        ("Breaking.Bad.S01E01.720p.BluRay.x264.mkv", "tv"),
-        ("Game of Thrones S01E01 1080p.mp4", "tv"),
-        ("Inception.2010.720p.BluRay.x264.mp4", "movie"),
-        ("Random.File.2020.1080p.mkv", "movie"),  # Default to movie
-    ],
-)
-def test_content_type_detection(filename, expected_type):
-    """Test content type detection."""
-    detected_type = detect_content_type(filename)
-    assert (
-        detected_type == expected_type
-    ), f"Content type mismatch: got '{detected_type}', expected '{expected_type}'"
+    assert info["season"] == expected_season, (
+        f"Season mismatch: got {info['season']}, expected {expected_season}"
+    )
+    assert info["episodes"] == expected_episodes, (
+        f"Episodes mismatch: got {info['episodes']}, expected {expected_episodes}"
+    )
+    assert info["year"] == expected_year, (
+        f"Year mismatch: got {info['year']}, expected {expected_year}"
+    )
+    assert info["resolution"] == expected_resolution, (
+        f"Resolution mismatch: got '{info['resolution']}', expected '{expected_resolution}'"
+    )
+    assert info["part"] == expected_part, (
+        f"Part mismatch: got {info['part']}, expected {expected_part}"
+    )
+    assert info["extra_type"] == expected_extra_type, (
+        f"Extra type mismatch: got {info['extra_type']}, expected {expected_extra_type}"
+    )
 
 
 @pytest.fixture
@@ -249,24 +236,20 @@ async def test_movie_organization(temp_dirs):
     """Test movie organization functionality."""
     source_dir, target_dir = temp_dirs
 
-    # Copy some test movies to source
     test_videos_dir = Path("test_videos/movies")
     if test_videos_dir.exists():
-        video_files = list(test_videos_dir.glob("*.mkv"))[:3]  # Test with first 3 files
+        video_files = list(test_videos_dir.glob("*.mkv"))[:3]
         for video_file in video_files:
             shutil.copy2(video_file, source_dir)
 
-        # Run organization
-        await organize_movies(str(source_dir), str(target_dir))
+        await organize(str(source_dir), str(target_dir), "movies", False)
 
-        # Check results
         source_files = list(source_dir.glob("*"))
         target_files = list(target_dir.glob("**/*"))
 
-        assert len(source_files) > 0, "Should have source files"
+        assert len(source_files) == 0, "Source should be empty after moving files"
         assert len(target_files) > 0, "Should have organized target files"
 
-        # Verify at least one organized file exists
         organized_files = [f for f in target_files if f.is_file()]
         assert len(organized_files) > 0, "Should have at least one organized file"
     else:
@@ -278,24 +261,20 @@ async def test_tv_organization(temp_dirs):
     """Test TV show organization functionality."""
     source_dir, target_dir = temp_dirs
 
-    # Copy some test TV shows to source
     test_videos_dir = Path("test_videos/tv_shows")
     if test_videos_dir.exists():
-        video_files = list(test_videos_dir.glob("*.mkv"))[:3]  # Test with first 3 files
+        video_files = list(test_videos_dir.glob("*.mkv"))[:3]
         for video_file in video_files:
             shutil.copy2(video_file, source_dir)
 
-        # Run organization
-        await organize_tv_shows(str(source_dir), str(target_dir))
+        await organize(str(source_dir), str(target_dir), "tv", False)
 
-        # Check results
         source_files = list(source_dir.glob("*"))
         target_files = list(target_dir.glob("**/*"))
 
-        assert len(source_files) > 0, "Should have source files"
+        assert len(source_files) == 0, "Source should be empty after moving files"
         assert len(target_files) > 0, "Should have organized target files"
 
-        # Verify season folders are created
         season_folders = [f for f in target_files if f.is_dir() and "Season" in f.name]
         assert len(season_folders) > 0, "Should have at least one season folder"
     else:
@@ -317,31 +296,19 @@ async def test_tv_organization(temp_dirs):
 )
 def test_edge_cases(filename):
     """Test edge case handling - ensure parsers don't crash on unusual filenames."""
-    # Test movie parsing doesn't crash
     try:
-        title, year, resolution, part, extra_type = parse_movie_info(filename)
-        assert title is not None, f"Movie parser should return a title for {filename}"
+        info = parse_media_info(filename, "movies")
+        assert info["title"] is not None, f"Parser should return a title for {filename}"
     except Exception as e:
         pytest.fail(f"Movie parsing failed for {filename}: {e}")
 
-    # Test TV parsing doesn't crash
     try:
-        series, season, episodes, year, resolution, part, extra_type = parse_tv_info(
-            filename
+        info = parse_media_info(filename, "tv")
+        assert info["title"] is not None, (
+            f"Parser should return a series for {filename}"
         )
-        assert series is not None, f"TV parser should return a series for {filename}"
     except Exception as e:
         pytest.fail(f"TV parsing failed for {filename}: {e}")
-
-    # Test content type detection doesn't crash
-    try:
-        content_type = detect_content_type(filename)
-        assert content_type in [
-            "movie",
-            "tv",
-        ], f"Content type should be 'movie' or 'tv', got {content_type}"
-    except Exception as e:
-        pytest.fail(f"Content type detection failed for {filename}: {e}")
 
 
 class TestExtensionHandling:
@@ -350,20 +317,15 @@ class TestExtensionHandling:
     def test_multiple_extensions_mr_robot(self):
         """Test handling of Mr. Robot style filenames with multiple extensions."""
         filename = "Mr.Robot.S01E01.eps1.0_hellofriend.mov.1080p.DTS-HD.MA.5.1.AVC.REMUX-FraMeSToR.mkv"
-        series, season, episodes, year, resolution, part, extra_type = parse_tv_info(
-            filename
-        )
+        info = parse_media_info(filename, "tv")
 
-        assert series == "Mr Robot"
-        assert season == 1
-        assert episodes == 1
-        # The key test: the extension handling should work properly
-        assert resolution == "1080p"  # Should still detect resolution
+        assert info["title"] == "Mr Robot"
+        assert info["season"] == 1
+        assert info["episodes"] == 1
+        assert info["resolution"] == "1080p"
 
     def test_multiple_extensions_preserve_episode_names(self):
         """Test that extensions in episode names are preserved."""
-        from core.tv_parser import get_base_filename_without_ext, get_real_extension
-
         filename = "Show.S01E01.episode.name.with.dots.mov.mkv"
 
         base = get_base_filename_without_ext(filename)
